Add Sentek sensor integration
Added the integration of Sentek sensordata and parsed the XML response to Java objects. The Sentek API endpoint and access token were added to the application.yml file for convenience. A parser in SentekSensorIntegrationService uses JAXB to transform XML responses from Sentek API into Java objects. Additionally, an AbstractIntegrationService was created to store common logic related to sensor integrations. Changed the name of AgvolutionSensorDataIntegrationServiceIT to AgvolutionUserIntegrationServiceIT for correctness. This feature was added to extend the range of sensor data the application can process.