Add functionality to fetch sensor data from Sentek API
Added a new method in SentekSensorIntegrationService to fetch all sensor data from the Sentek API. This results in a list of Logger objects after successful parsing of the XML response. In case of any error during the fetching process, a relevant error is now thrown. Also renamed a method in the AbstractIntegrationService to align with its functionality and added a new test case in SentekSensorIntegrationServiceTest ensuring the API fetch is functioning correctly. These changes allow the application to extend its integration capability with Sentek sensors.