Move Sentek integration models to a xml package and add CSV parsing for sensor data.
Models specific for XML integration with Sentek systems were moved to their own package de.app.fivegla.integration.sentek.model.xml from the generic model package for clarity and better organization. This change makes it easier to understand the different types of models the application is using and can help to avoid any confusion when dealing with different representations of the same entity for different integrations.