Feature/enable access to the swagger page (#59)
* Refactor routes and update API key authentication

The change includes the addition of a new base mapping '/api/v1' which is now prefixed to all existing routes. Also, within the `ApiKeyInterceptor`, checks were added to ensure that API key authentication is required only for API requests, improving app's security. Added 'Slf4j' annotation to enable logging.

* Update version in pom.xml.

The version of the artifact 'sensor-integration-services' in the pom.xml file was updated from 6.0.0 to 6.1.0. This update was necessary to reflect the changes made in the application as a result of the refactoring of routes and enhancing the API key authentication.