Add Fiware status check endpoint and exception handling.
Implementing a new endpoint `/info/fiware` in the `InfoController` to check the status of the Fiware connection. This includes updates to `Application.java` for dependency injection of the `StatusService`, creation of a new DTO `FiwareStatusResponse`, and an update to the `CustomExceptionHandler` to handle `FiwareIntegrationLayerException`. Also, the version of 'fiware-integration-layer' was increased in the `pom.xml` reflecting the recent changes in the Fiware Integration Layer.