Feature/add api call to check for fiware status (#61)
* Add Fiware status check endpoint and exception handling.

Implementing a new endpoint `/info/fiware` in the `InfoController` to check the status of the Fiware connection. This includes updates to `Application.java` for dependency injection of the `StatusService`, creation of a new DTO `FiwareStatusResponse`, and an update to the `CustomExceptionHandler` to handle `FiwareIntegrationLayerException`. Also, the version of 'fiware-integration-layer' was increased in the `pom.xml` reflecting the recent changes in the Fiware Integration Layer.

* Update version in pom.xml for new features

The version in pom.xml was increased from 6.1.0 to 6.2.0. This update is to reflect the recent changes in the Fiware Integration Layer which includes the implementation of a new endpoint to check the status of the Fiware connection and enhancement of exception handling.