Introduce fiware subscription configuration
The configuration for fiware subscription is now introduced, allowing to enable or disable subscriptions. This adds flexibility and control over when to receive notifications from FIWARE. This modification involved changes in both test and main application.yml files, updating property injection in Application.java and DataImportEventHandler.java. It's also included in additional-spring-configuration-metadata.json file for description on this new implementation.