Update version in pom.xml from 6.2.0 to 6.3.0
An upgrade was needed to the project's version in the pom.xml file in order to reflect the changes made for the addition of Fiware subscription configuration. Updating the version number in pom.xml helps to track that a new feature has been introduced in this update and is recommended for good version control.