Feature/subscribe to 3rd party before import (#62)
* Add notification URL and include SubscriptionService

Added a new notification URL to the application.yml configuration file and updated Application.java to use this URL. A SubscriptionService bean was created which will use the notification URL. The SubscriptionService was also added to the DataImportEventHandler which will handle the subscription and reset process. Furthermore, the version of fiware-integration-layer in pom.xml was bumped up to 5.3.0 to reflect these new changes and improvements.

* Introduce fiware subscription configuration

The configuration for fiware subscription is now introduced, allowing to enable or disable subscriptions. This adds flexibility and control over when to receive notifications from FIWARE. This modification involved changes in both test and main application.yml files, updating property injection in Application.java and DataImportEventHandler.java. It's also included in additional-spring-configuration-metadata.json file for description on this new implementation.

* Update version in pom.xml from 6.2.0 to 6.3.0

An upgrade was needed to the project's version in the pom.xml file in order to reflect the changes made for the addition of Fiware subscription configuration. Updating the version number in pom.xml helps to track that a new feature has been introduced in this update and is recommended for good version control.