Modify scheduled task timings and data import in application.yml
Updated the schedule of task execution in the application.yml of src/test/resources and src/main/resources, changing it from every 15 minutes to every hour. Also reduced the days in the past for initial data import from 7 to 1 in the testing environment to optimize test resources. These changes aim to optimize system resources and increase efficiency.