Bugfix/fix cron expression (#66)
* Modify scheduled task timings and data import in application.yml

Updated the schedule of task execution in the application.yml of src/test/resources and src/main/resources, changing it from every 15 minutes to every hour. Also reduced the days in the past for initial data import from 7 to 1 in the testing environment to optimize test resources. These changes aim to optimize system resources and increase efficiency.

* Update version number in pom.xml

Bumped up the version number of sensor-integration-services in pom.xml from 6.4.0 to 6.4.1. This was necessary to signal a new release, incorporating minor bug fixes and system improvements.