Feature/add custom context path to support ingress (#69)
* Set context-path for application.yml

Added context-path /5gla-sis in the application.yml file. This provides a proper endpoint for accessing information regarding the system's health, info, and prometheus. This makes it easier for network administrators to monitor the status of the system.

* Update pom.xml version number

Updated the version number from 6.5.0 to 7.0.0 in pom.xml. This change denotes a major update and accurately reflects the current state of the application as per Semantic Versioning rules.