Add CORS settings to application configuration
A new CORS configuration was implemented in the application.yml file under the 'subscriptions' section. The addition of 'allowed-origins' will enable control over the origins that are able to access the application's resources, enhancing overall security. The origins permitted are set to 'http://localhost:8080' and 'https://app.5gla.de/'. They can be adjusted via the 'CORS_ALLOWED_ORIGINS' environment variable.