Feature/fix cors problems for cluster config (#70)
* Add CORS configuration and allowed origins setting

Introduced CORS configuration and added 'allowed-origins' setting to the application configuration. CORS configuration allows the application to control which origins can access the resources. This is a necessary step to secure the application. 'Allowed-origins' can be set via environment variable 'CORS_ALLOWED_ORIGINS'.

* Add CORS settings to application configuration

A new CORS configuration was implemented in the application.yml file under the 'subscriptions' section. The addition of 'allowed-origins' will enable control over the origins that are able to access the application's resources, enhancing overall security. The origins permitted are set to 'http://localhost:8080' and 'https://app.5gla.de/'. They can be adjusted via the 'CORS_ALLOWED_ORIGINS' environment variable.

* Update CORS mappings in application configuration

Added @SuppressWarnings annotation to the CORS configuration method. This change is set to prevent warnings from the Nullable Problems of the IDE and enhance readability of the code. This update targets the addCorsMappings method enabling control over the origins that can interact with the application's API.