Feature/externalize context path (#71)
* Update CORS mappings in application configuration

Added @SuppressWarnings annotation to the CORS configuration method. This change is set to prevent warnings from the Nullable Problems of the IDE and enhance readability of the code. This update targets the addCorsMappings method enabling control over the origins that can interact with the application's API.

* Update project version in pom.xml

Incremented the project version from 7.0.0 to 7.0.2 in the pom.xml file. This update is necessary to reflect the recent changes and improvements made in the project's codebase. It enables correct versioning and organization of the project's future releases.

* Add exception handling for Fiware Integration

This update adds error handling for the Fiware Integration part of the application. This was necessary as there were cases where errors in the integration service were causing unhandled exceptions and potentially hindering data persisting process. Now handling FiwareIntegrationLayerException in all necessary services: Farm21, SoilScout, Agvolution, Weenat, Sensoterra. This will ensure the application's robustness by enabling it to handle and log these exceptions properly before interrupting the entire flow.