Feature/update swagger descriptions for real servers (#72)
* Add server annotations for Swagger in Application

Server annotations were introduced to Swagger in Application to demarcate Production and Development servers distinctly. This makes it easier for their endpoints to be accessed accordingly based on specific server needs.

* Update pom.xml version for sensor-integration-services

The version of sensor-integration-services in pom.xml was updated from 7.0.2 to 7.0.3. This change was necessary as part of server annotations for Swagger in Application. It aids in distinguishing between Production and Development servers, allowing distinct access to their respective endpoints based on specific server needs.