Feature/fix server url (#73)
* Update Production server URL in Application.java

Changed the URL for the production server in the Application.java file. The updated URL now includes the path to the sensor-integration-services, making it easier for the clients to locate the correct endpoint. This change provides better direction and organization for accessing different components of the 5gla application.

* Update Production server URL in Application.java

Changed the URL for the production server in the Application.java file. The updated URL now includes the path to the sensor-integration-services, making it easier for the clients to locate the correct endpoint. This change provides better direction and organization for accessing different components of the 5gla application.