Feature/make imports fail safe (#74)
* Update exception handling in service wrapper classes

Changed the specific exception FiwareIntegrationLayerException to a broader RuntimeException in the catch blocks across all FiwareIntegration service wrapper classes. This change is aimed to handle a wider range of unexpected runtime errors, not just specific integration layer exceptions, providing robust error handling and improving application stability.

* Update project version in pom.xml

Bumped up the version of the artifact 'sensor-integration-services' from 7.0.4 to 7.0.5 in pom.xml file. This is part of the routine version update to reflect the latest changes and improvements made in the application, ensuring the new build carries the incremented version.

* Update project version in pom.xml

Bumped up the version of the artifact 'sensor-integration-services' from 7.0.4 to 7.0.5 in pom.xml file. This is part of the routine version update to reflect the latest changes and improvements made in the application, ensuring the new build carries the incremented version.