Adjust API URL and rename project in several files
The URL in the Application.java was updated to reflect the new API endpoint, from "/5gla-sis" to "/api". Similarly, the project name was changed from 'sensor-integration-services' to 'api' in both pom.xml and docker-compose.yml files. All these changes are made to align with the evolution of the project structure - 'sensor-integration-services' doesn't reflect the broader functionality of the project anymore so it's renamed to 'api'.