Feature/update module name and server path (#75)
* Adjust API URL and rename project in several files

The URL in the Application.java was updated to reflect the new API endpoint, from "/5gla-sis" to "/api". Similarly, the project name was changed from 'sensor-integration-services' to 'api' in both pom.xml and docker-compose.yml files. All these changes are made to align with the evolution of the project structure - 'sensor-integration-services' doesn't reflect the broader functionality of the project anymore so it's renamed to 'api'.

* Update project version in pom.xml

Project version was updated from 7.0.5 to 8.0.0 in pom.xml. The change is part of the ongoing preparation for a new major release, hence the shift from version 7.x to version 8.x. This update reflects significant enhancements and potential changes in the application's functionality.