Feature/adapt documentation in GitHub (#76)
* Update README with detailed API and setup info

The README.md file has been updated to provide a more comprehensive description of the 5gLA project, the integration service’s role in it, the source code's location, and how to set up and run the application within an IDE. The previous brief explanation about running the service with different profiles didn't give enough context or technical details for new users. This change aims to better guide users in setting up and understanding the application.

* Updated openapi.json to improve API description and expand endpoints

Significantly updated openapi.json to provide more comprehensive information about the API, including its purpose, contact details, terms of service, and licensing details. The version has also been updated from 4.0.0 to 8.0.0. Many new endpoints have been added to cover upcoming functionality related to handling and processing sensor data from various devices. Moreover, more detailed schemas for requests and responses have been provided. This updated API documentation will enable developers and users to better understand and interact with the service's expanded functionality.

* Update API metadata in Application.java

Changed the API description and title in Application.java file to better reflect the focus of working with the 5gLa platform. The new description provides a more concise overview of the 5gLa project and the functionalities of the integration service. This supports a clearer understanding of the service's purpose and scope for any new developers or users.