Feature/fail safe subscription handling (#77)
* Handle FiwareIntegrationLayerException in DataImportEventHandler

Extended subscriptionService.subscribeAndReset with a try/catch block in the handleDataImportEvent function in DataImportEventHandler.java. This update manages potential FiwareIntegrationLayerExceptions that can occur while subscribing to device measurement notifications. Without this update, the system could crash if such exceptions were thrown, leading to user inconvenience.

* Update version in pom.xml

Updated the version of the application in the pom.xml file from 8.0.0 to 8.1.0. This step is necessary for the new release, which includes additional exception handling code in DataImportEventHandler for managing potential FiwareIntegrationLayerExceptions, as these could previously lead to system crashes.