Update version number in pom.xml file
Version number in pom.xml file has been updated from 8.1.0 to 8.2.0. This was done to reflect the new changes in the project related to exception handling in Farm21 integration services. This marks the readiness of the new release with improved stability and robustness in exception handling when interacting with the Farm21 API.