Feature/make calls to 3rd party apis failsafe (#78)
* Add exception handling to Farm21 integration services

Added try-catch blocks in Farm21SensorDataIntegrationService and Farm21SensorIntegrationService to better handle potential runtime exceptions. This change was necessary to prevent potential system crashes due to unhandled exceptions when fetching data from the Farm21 API, thus improving the robustness of the system. Also, added a new error code "FARM21_COULD_NOT_FETCH_DEVICE_DATA" to handle fetching device data errors.

* Update version number in pom.xml file

Version number in pom.xml file has been updated from 8.1.0 to 8.2.0. This was done to reflect the new changes in the project related to exception handling in Farm21 integration services. This marks the readiness of the new release with improved stability and robustness in exception handling when interacting with the Farm21 API.