Feature/update url for swagger UI (#79)
* Refactor error message and update API url

Adjusted the error message in DataImportEventHandler to better reflect the underlying issue when a subscription to device measurement notifications cannot be established. Additionally, the API url in the Application file has been updated to ensure it points to the correct endpoint. Both changes are aimed at improving the clarity and maintainability of the codebase.

* Update version in pom.xml

Version in pom.xml file updated from 8.2.0 to 8.2.1. This change has been made to keep the project versioning consistent and up-to-date as some modifications in the error message and API URL have been made in the project.