Add DeviceRegistrationController for device registration
Added a DeviceRegistrationController to provide RESTful endpoints for device registration. This controller will handle all the interactions related to the registration of the devices in the system. To accommodate this, DEVICE_REGISTRATION mapping was added in BaseMappings class and DEVICE_REGISTRATION tag was added in OperationTags class. The addition of these features will improve device management and enable more efficient interaction with registered devices.