Update fiware-integration-layer version in pom.xml
The 'fiware-integration-layer' dependency version was upgraded from 6.0.0 to 6.1.0 in pom.xml. This change was necessary to make use of the new features offered in the latest version. Additionally, several changes were made to improve device registration monitoring to help track and optimize the performance of device registration processes in future updates.