Update DeviceRegistrationController and add DeviceRegistrationRequest class
Included DeviceRegistrationRequest object in DeviceRegistrationController's register method to capture registration details. A new DeviceRegistrationRequest DTO is introduced explicitly for a device registration process to accept the manufacturer, ID and location details (latitude and longitude) of the device.

Adjusted DeviceRegistrationController to use this DTO in its register method. This improvement makes the device registration more flexible, maintaining the separation of concerns and cleaner handling of registration details. It also improves readability and maintainability of the code.