Update API version and add descriptions to OpenAPI
Updated the API version from `9.1.1` to `10.0.0` within the OpenAPI JSON structure. Added descriptions to request parameters, such as `droneId`, `transactionId`, `sensorId`, `plotId`, and `deviceId`, to make the API more understandable for the users. Modified the request and response content type to `application/json` from `*/*` for standardization. Renamed a number of paths including `/v1/data-logging` to `/v1/device`, and associated tags and operation ids to reflect the new logic. Provided new endpoints to register and delete a device, and described responses for them. Plus, added further descriptions and properties to the request schemas. These changes are implemented for cleanly structuring the OpenAPI, making API more user-friendly and maintaining consistency throughout the schema.