Add support for GeoJSON in AgriCrop service
This commit introduces the ability to import and parse GeoJSON information in the `AgriCropService`. A new `AgriCropController` was created to handle the importing of GeoJSON and a new endpoint, `/agri-crop/geo-json` was added. The `gt-geojson` dependency was added to `pom.xml` to facilitate the handling of GeoJSON. This optimization was necessary for the introduction of new features to work with GeoJSON data formats within agricultural data scopes. The "storage-directory" value in `application.yml` has also been fixed by removing an extra '1'.

Additionally, an error definition was added for when the parsing of GeoJSON fails and a test to ensure the correct operation of parsing GeoJSON has been provided.