Add Slf4j logging and modify GeoJSON parsing method
This commit includes changes to introduce logging via Slf4j and renaming the method parse to parseFeature in the AgriCropService. The reason for adding logging is to enhance troubleshooting capabilities in the application. The renaming of the parse method is carried out to emphasize the parsing of a single GeoJSON feature, which brings more clarity to the code. Changes in the AgriCropController and AgriCropServiceTest were consequently made to cater for the new method name.