Add suppress warning annotation to AgriCropServiceTest file
The code has been updated to add the suppress warning annotation ("FieldCanBeLocal") to the AgriCropServiceTest.java file. This was done because the variable 'feature' needs to be accessed across the different test methods and cannot be made local, although the IDE suggests so. This will prevent inappropriate warning messages from being generated.