Update scheduled task configuration
Introduced initial delay and fixed delay configuration attributes for the data import scheduled task. The previously used cron expression was replaced to provide more flexibility for the scheduling. It allows for a delay before the first execution (initial-delay), and a fixed delay between the completion of one execution and the start of another. Both are now configured in the "application.yml" file in ISO8601 duration format. Also, these new configurations were added to the metadata file for usability.