Update data import schedule in application.yml
The scheduling properties for data import in application.yml were changed, to provide more control over the timing. Instead of using a cron expression to schedule data import every hour, two separate properties, 'initial-delay' and 'delay' are introduced with ISO-8601 durations. Now, the initial delay is 10 seconds, with a recurring delay of 30 minutes. This is expected to optimize data import process and align it better with the requirements.