Added QA and DEV server details and updated version in API
In this commit, we have added details of QA and DEV servers to the Application.java file, this is needed for better environment-oriented testing and development efficiency. Also, the 'version' value is updated in openapi.json to reflect the latest API changes. Furthermore, expanded the API's capabilities by adding 'agri-crop' importing functions along with detailed descriptions and response scenarios.