Implement job disabling feature for specific manufacturer
This update introduces the functionality to disable certain jobs for specific manufacturers. The feature includes addition of a new 'DisabledJob' class to facilitate this. Methods to disable a job and to check whether a job is disabled for a given manufacturer have been added to the 'ApplicationData' and 'ApplicationDataRepository' classes.