Update device measurements in Fiware integration
The device measurements handling in Fiware integration service have been significantly updated. The FiwareIdGenerator class is removed, and measurement creation is migrated to use the DeviceMeasurementBuilder. Also, Fiware integration version is updated in pom.xml, and the MeasurementType enum is added. Adjustments are made in Application.java and SentekFiwareIntegrationServiceWrapper.java accordingly to accommodate these changes.