Add check for duplicate Tenant IDs during creation
This commit introduces a validation check during the creation of a new 'Tenant'. The system will now prevent the creation of a new Tenant if another with the same ID already exists. Corresponding updates were made in TenantService, ApplicationDataRepository, and Error classes. Minor changes were also made in the .gitignore file and the CreateTenantResponse class.