Replace MicaSense references with generic Image
In the ApplicationDataRepository and related classes, references to `MicaSenseImage` and related classes have been replaced with more generic classes. This allows for more flexibility in the types of file formats the application can handle, with `Image` and `ImageChannel` usable in a wider range of situations. The renaming was performed across several files to maintain consistency.