Update MeasurementType and rename methods in repositories
This commit modifies the MeasurementType enumeration to include a key for each enum value, affecting how the MeasurementType is used across various services. It also changes a method name from 'findAllTenants' to 'findTenants' in the ApplicationDataRepository and reflects this change wherever the method is called. The commit further adds a subscription initialization on application's startup in SubscriptionStatus.