Rename enum "MeasurementType" to "EntityType"
The "MeasurementType" enum has been renamed to "EntityType" across various classes. This also meant updating references in other classes like the controller and service wrappers. The purpose of this change is to have a more accurate name for this part of the system. As a result of this change, an "AgriCrop" enum was also added.