Implement ZoneOrDefaultValue in service wrappers
The measure is to ensure default values are used for the “zone” field in different services if no zone is provided. Added the ZoneOrDefaultValue class and updated all service wrappers to use this class for zone fields. For each entity where a zone value is applicable, if no zone value is provided, a default value will be assigned.