Add functionality to create default group for tenants
Removed the 'defaultGroups' list from the 'ApplicationData' class and added functionality to create a default group for each tenant. A new event 'CreateDefaultGroupForTenantEvent' is published when a new tenant is created in the 'TenantService', which is handled by the new 'CreateDefaultGroupForTenantEventHandler'. The 'GroupRepository' class has been updated to include a 'createDefaultGroup' method.