Refactor Group creation to GroupService
The responsibility of creating a default group for a tenant has been moved to GroupService from GroupRepository. All references to this method throughout the project have also been updated. This better aligns with the principles of Separation of Concerns, with business logic now located within a service class rather than a repository.