Refactor data import logic to deal with missing group association
The commit introduces changes to handle the cases where a sensor does not have a group associated with it. The data import functions were modified to no longer require a Group argument, and a mechanism was added in the GroupService to find a group by the tenant and sensor ID. In cases where no group is found, the default group for the tenant is used. A warning is logged if the group for a sensor is the default one.